Written Game class beside move validation.

+++ sources/Point.hpp
@@ -8,10 +8,10 @@
 
 struct Point {
 
-    Point(uint8_t x, uint8_t y): x_(x), y_(y) { }
+    Point(int x, int y): x_(x), y_(y) { }
 
-    uint8_t x_;
-    uint8_t y_;
+    int x_;
+    int y_;
 
 };
